chore: 음성 녹음 훅에서 타이머 관련 내용을 훅으로 분리 #36
